[feat] boardDTO 전환 (#57)
* dto, json 구현

* 예외처리 진행중

* save

* [feat]board DTO, 예외처리,통합테스트완료

---------

+++ src/main/java/com/example/recruit_page_wwy/board/BoardService.java
@@ -1,6 +1,8 @@
 package com.example.recruit_page_wwy.board;
 
 
+import com.example.recruit_page_wwy._core.error.ex.ExceptionApi403;
+import com.example.recruit_page_wwy._core.error.ex.ExceptionApi404;
 import com.example.recruit_page_wwy.reply.Reply;
 import com.example.recruit_page_wwy.reply.ReplyRepository;
 import com.example.recruit_page_wwy.user.User;
@@ -12,6 +14,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
 @RequiredArgsConstructor
 @Service
@@ -26,8 +29,10 @@ public class BoardService {
     // TODO : save -> persist
     // TODO : 저장 후 DTO에 담아서 반환
     @Transactional
-    public void boardSave(BoardRequest.SaveDTO saveDTO) {
-        boardRepository.save(saveDTO.getUser_id(), saveDTO.getTitle(), saveDTO.getContent());
+    public BoardResponse.DTO boardSave(BoardRequest.SaveDTO reqDTO, User sessionUser) {
+        Board board = reqDTO.toEntity(sessionUser);
+        Board boardPS = boardRepository.save(board);
+        return new BoardResponse.DTO(boardPS);
     }
 
     public BoardResponse.ListDTO boardList(Integer page, User sessionUser, String keyword) {
@@ -53,12 +58,33 @@ public class BoardService {
     // TODO : update -> dirty checking
     // TODO : 업데이트 후 DTO에 담아서 반환
     @Transactional
-    public void boardUpdate(Integer id, BoardRequest.UpdateDTO updateDTO) {
-        boardRepository.boardUpdate(id, updateDTO.getTitle(), updateDTO.getContent());
+    public BoardResponse.DTO boardUpdate(Integer id, BoardRequest.UpdateDTO reqDTO, Integer sessionUserId) {
+        Board board = boardRepository.findById(id);
+
+        System.out.println("sessionUserId = " + sessionUserId);
+        if (board == null) {
+            throw new ExceptionApi404("자원을 찾을 수 없습니다");
+        }
+
+        if (!Objects.equals(board.getUser().getId(), sessionUserId)) {
+            throw new ExceptionApi403("권한이 없습니다");
+        }
+
+        board.update(reqDTO);
+
+        return new BoardResponse.DTO(board);
     }
 
     @Transactional
-    public void boardDelete(Integer id) {
+    public void boardDelete(Integer id, Integer sessionUserId) {
+        Board boardPS = boardRepository.findById(id);
+        if (boardPS == null) {
+            throw new ExceptionApi404("자원을 찾을 수 없습니다");
+        }
+
+        if (!Objects.equals(boardPS.getUser().getId(), sessionUserId)) {
+            throw new ExceptionApi403("권한이 없습니다");
+        }
         boardRepository.delete(id);
     }
